# src/server/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Cyberacy API",
    "description": "Documentation for Cyberacy API."
  },
  "host": "cyberacy.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Rentrez un token valide dans le format **Bearer &lt;token>**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/manifestation": {
      "post": {
        "tags": [
          "Manifestation"
        ],
        "description": "Ajouter une nouvelle manifestation.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "Manifestation"
        ],
        "description": "Récupération des manifestations.",
        "parameters": [
          {
            "name": "mine",
            "in": "query",
            "description": "Voir uniquement les manifestation auxquelles je participe.",
            "type": "boolean"
          },
          {
            "name": "includeAborted",
            "in": "query",
            "description": "Inclure les manifestation annuler.",
            "type": "boolean"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manifestation/aborted": {
      "patch": {
        "tags": [
          "Manifestation"
        ],
        "description": "Annuler une manifestation existante.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "reason": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manifestation/{id}/participate": {
      "post": {
        "tags": [
          "Manifestation"
        ],
        "description": "Participer à une manifestations.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "parameter",
            "required": true,
            "type": "number",
            "description": "Id de la manifestation."
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manifestation/option": {
      "post": {
        "tags": [
          "Manifestation"
        ],
        "description": "Ajouter une option à une manifestation.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manifestation/{id}/option": {
      "get": {
        "tags": [
          "Manifestation"
        ],
        "description": "Récupérer les options de une manifestation.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manifestation/option/{id}": {
      "delete": {
        "tags": [
          "Manifestation"
        ],
        "description": "Supprimer une option de une manifestation.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manifestation/step": {
      "post": {
        "tags": [
          "Manifestation"
        ],
        "description": "Ajouter une étape à une manifestation.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manifestation/{id}/step": {
      "get": {
        "tags": [
          "Manifestation"
        ],
        "description": "Récupérer les étapes d\\'une manifestation.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/manifestation/step/{id}": {
      "delete": {
        "tags": [
          "Manifestation"
        ],
        "description": "Supprimer une étape de une manifestation.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Authentification"
        ],
        "description": "Créer un compte basé sur le NIR (numéro de sécurité sociale).",
        "parameters": [],
        "responses": {}
      }
    },
    "/login_app": {
      "post": {
        "tags": [
          "Authentification"
        ],
        "description": "Connexion et retourne le Bearer token (pour application Android).",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nir": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/login_stats": {
      "post": {
        "tags": [
          "Authentification"
        ],
        "description": "Connexion et retourne le Bearer token (pour application de statistiques IOS).",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nir": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/login_bo": {
      "post": {
        "tags": [
          "Authentification"
        ],
        "description": "Connexion et retourne le Bearer token (pour le back-office Flutter).",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nir": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/type_step": {
      "get": {
        "tags": [
          "References"
        ],
        "description": "Récupération des types d",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/sex": {
      "get": {
        "tags": [
          "References"
        ],
        "description": "Récupération des civilités.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/type_vote": {
      "get": {
        "tags": [
          "References"
        ],
        "description": "Récupération des types de vote.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/political_edge": {
      "get": {
        "tags": [
          "References"
        ],
        "description": "Récupération des bords politiques.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/region": {
      "get": {
        "tags": [
          "Geography"
        ],
        "description": "Récupération des régions.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/region/{code_insee}/department": {
      "get": {
        "tags": [
          "Geography"
        ],
        "description": "Récupération des départements par régions.",
        "parameters": [
          {
            "name": "code_insee",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/department": {
      "get": {
        "tags": [
          "Geography"
        ],
        "description": "Récupération des départements.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/department/{code}/town": {
      "get": {
        "tags": [
          "Geography"
        ],
        "description": "Récupération des communes par département.",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/town": {
      "get": {
        "tags": [
          "Geography"
        ],
        "description": "Récupération des communes.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/vote": {
      "post": {
        "tags": [
          "Vote"
        ],
        "description": "Ajout de un nouveau vote.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/political_party": {
      "post": {
        "tags": [
          "Political party"
        ],
        "description": "Ajout de un nouveau parti politique si il éxiste dans la BDD de INSEE.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "Political party"
        ],
        "description": "Récupération des partis politiques.",
        "parameters": [
          {
            "name": "siren",
            "in": "query",
            "description": "Recherher par le SIREN.",
            "type": "string"
          },
          {
            "name": "mine",
            "in": "query",
            "description": "Recherher les partie auxquel on a adhéré.",
            "type": "boolean"
          },
          {
            "name": "includeLeft",
            "in": "query",
            "type": "string"
          },
          {
            "name": "idPoliticalParty",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/political_party/{id}/join": {
      "post": {
        "tags": [
          "Political party"
        ],
        "description": "Adhérer à un parti politique.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/political_party/left": {
      "delete": {
        "tags": [
          "Political party"
        ],
        "description": "Partir d'un parti politique.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/political_party/annual_fee": {
      "post": {
        "tags": [
          "Political party"
        ],
        "description": "Ajoute une cotisation annuelle.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/political_party/{id}/annual_fee": {
      "get": {
        "tags": [
          "Political party"
        ],
        "description": "Récupère les cotisations annuelles d",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/political_party/upload_logo": {
      "post": {
        "tags": [
          "Political party"
        ],
        "description": "Upload le logo du parti politique.",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/meeting": {
      "post": {
        "tags": [
          "Meeting"
        ],
        "description": "Ajoute un nouveau meeting.",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "Meeting"
        ],
        "description": "Récupération des meetings.",
        "parameters": [
          {
            "name": "mine",
            "in": "query",
            "description": "Voir uniquement les meetings auxquelles je participe.",
            "type": "boolean"
          },
          {
            "name": "town",
            "in": "query",
            "type": "string"
          },
          {
            "name": "idPoliticalParty",
            "in": "query",
            "type": "string"
          },
          {
            "name": "includeAborted",
            "in": "query",
            "description": "Inclure les meetings annuler.",
            "type": "boolean"
          },
          {
            "name": "includeCompleted",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/meeting/aborted": {
      "patch": {
        "tags": [
          "Meeting"
        ],
        "description": "Annuler un meeting existant.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "reason": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/meeting/participate/{id}": {
      "post": {
        "tags": [
          "Meeting"
        ],
        "description": "Participer à un meeting.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Meeting"
        ],
        "description": "Annuler sa participation à un meeting.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "reason",
            "in": "query",
            "description": "La raison de l annulation",
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  }
}